+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - **_Breaking Change_**: stricter trait bound for `MockLambdaRuntimeApiServer` methods.
 - **_Breaking Change_**: `MockLambdaRuntimeApiServer::passthrough` will create a new client every time.
 - **_Breaking Change_**: remove `RunningProxy::client`. `RunningProxy` and `Proxy::spawn` doesn't have generic params anymore.
+- **_Breaking Change_**: rename `LambdaRuntimeApiClient::start` to `new`.
 
 ## v0.1.1